Crate build_info_common

Common types used by the build-info and build-info-build crates.

User code should not depend on this crate directly, but rather depend on build-info (as a [dependency]) and build-info-build (as a [build-dependency]). The types provided herein are reexported by build-info and should be used that way. For example, build_info_common::BuildInfo should be used as build_info::BuildInfo instead.

Re-exports

pub use chrono;

pub use semver;

Structs

BuildInfo

Information about the current build

CompilerInfo

rustc version and configuration

CpuInfo

CrateInfo

Information about the current crate (i.e., the crate for which build information has been generated)

GitInfo

Information about a git repository

TargetInfo

VersionedString

Used internally to ensure that build-info and build-info-build use the same version of build-info-common.

Enums

CompilerChannel

rustc distribution channel (some compiler features are only available on specific channels)

Endianness

CPU Endianness

OptimizationLevel

The various possible optimization levels

VersionControl

Support for different version control systems

Functions

crate_version

Gets the version of the build-info-common crate (this crate)